feat: setup resource title & extract necessary helpers
